add project function

+++ screens/ProjectListScreen.js
@@ -8,12 +8,12 @@ import * as firebase from 'firebase';
 let db;
 let currentUser;
 
-export default class ContactListScreen extends React.Component {
+export default class ProjectListScreen extends React.Component {
 	constructor(props) {
 		super(props);
 		this.ds = new ListView.DataSource({ rowHasChanged: (r1, r2) => r1 !== r2 });
 		this.state = {
-			contacts: [],
+			projects: [],
 			loading: true,
 		};
 		currentUser = firebase.auth().currentUser;
@@ -27,49 +27,49 @@ export default class ContactListScreen extends React.Component {
 	}
 
 	componentDidMount() {
-		this.unsubscribe = db.collection('contacts').where(this.replaceAll(currentUser.email, '.', '%2E'), '==', true)
+		this.unsubscribe = db.collection('projects').where(this.replaceAll(currentUser.email, '.', '%2E'), '==', true)
 			.onSnapshot((snapShot) => {
-				this.setState({ contacts: [], loading: true });
+				this.setState({ projects: [], loading: true });
 				if (snapShot.empty) {
 					this.setState({ loading: false });
 				} else {
 					snapShot.docs.forEach((doc, i) => {
-						const docData = Object.keys(doc.data());
-						// const result = await docData.filter(e => currentUser.email);
-						docData.forEach((userEmail) => {
-							const replacedEmail = this.replaceAll(userEmail, '%2E', '.');
-							if (replacedEmail !== currentUser.email) {
-								db.collection('users').where('email', '==', replacedEmail)
-									.get()
-									.then((userDocs) => {
-										if (snapShot.docs.length <= i+1) {
-											this.setState({ contacts: [...this.state.contacts, { ...userDocs.docs[0].data(), id: doc.id }], loading: false });
-										} else {
-											this.setState({ contacts: [...this.state.contacts, { ...userDocs.docs[0].data(), id: doc.id }]});
-										}
-									});
-							}
-						});
+						if (snapShot.docs.length <= i + 1) {
+							this.setState({ projects: [...this.state.projects, { ...doc.data(), id: doc.id }], loading: false });
+						} else {
+							this.setState({ projects: [...this.state.projects, { ...doc.data(), id: doc.id }] });
+						}
 					});
 				}
 			});
 	}
-
+    
 	componentWillUnmount() {
 		this.unsubscribe();
 	}
 
 	deleteRow(data, secId, rowId, rowMap) {
-		rowMap[`${secId}${rowId}`].props.closeRow();
-		db.collection('contacts').doc(data.id)
-			.delete()
-			.then(() => {
-				const contacts = this.state.contacts.filter(e => e.id !== data.id);
-				this.setState({contacts});
-			})
-			.catch(error => {
-				alert(error);
-			});
+		Alert.alert('確認', '本当に削除しますか？関連するスケジュールもすべて削除されます', [{
+			text: 'はい', onPress: () => {  
+				rowMap[`${secId}${rowId}`].props.closeRow();
+				db.collection('projects').doc(data.id)
+					.delete()
+					.then(() => {
+						const projects = this.state.projects.filter(e => e.id !== data.id);
+						this.setState({ projects });
+						db.collection('schedules').where(data.id, '==', true)
+							.get()
+							.then((snapShot) => {
+								snapShot.forEach(doc => {
+									doc.ref.delete();
+								});
+							});
+					})
+					.catch(error => {
+						alert(error);
+					});
+			}
+		}, { text: 'キャンセル' }]); 
 	}
 
 	render() {
@@ -82,7 +82,7 @@ export default class ContactListScreen extends React.Component {
 							} />
 						</Left>
 						<Body>
-							<Title>コンタクト</Title>
+							<Title>プロジェクト</Title>
 						</Body>
 					</Header>
 					<LoadingScreen />
@@ -97,7 +97,7 @@ export default class ContactListScreen extends React.Component {
 							} />
 						</Left>
 						<Body>
-							<Title>コンタクト</Title>
+							<Title>プロジェクト</Title>
 						</Body>
 					</Header>
 					<View style={{ flex: 1 }}>
@@ -105,20 +105,15 @@ export default class ContactListScreen extends React.Component {
 							<List
 								leftOpenValue={75}
 								rightOpenValue={-75}
-								dataSource={this.ds.cloneWithRows(this.state.contacts)}
+								dataSource={this.ds.cloneWithRows(this.state.projects)}
 								renderRow={data =>
-									<ListItem avatar>
-										<Left>
-											<Thumbnail small style={{ marginLeft: '5%' }} source={{ uri: data.image ? data.image : 'https://firebasestorage.googleapis.com/v0/b/rehearsalplanner-f7b28.appspot.com/o/%E3%83%87%E3%83%95%E3%82%A9%E3%83%AB%E3%83%88%E3%82%A2%E3%82%A4%E3%82%B3%E3%83%B3.png?alt=media&token=70a6f7fd-a9b0-4790-a4a8-58c3d94823c3' }} />
-										</Left>
+									<ListItem onPress={() => this.props.navigation.navigate('ProjectScheduleListScreen', { project: data })}>
 										<Body>
 											<Text>{data.name}</Text>
-											<Text note>{data.email}</Text>
 										</Body>
 									</ListItem>}
 								renderLeftHiddenRow={data =>
-									<Button full onPress={() => {
-									}}>
+									<Button full onPress={() => this.props.navigation.navigate('ProjectEditScreen', { project: data })}>
 										<Icon active name="information-circle" />
 									</Button>}
 								renderRightHiddenRow={(data, secId, rowId, rowMap) =>
@@ -131,7 +126,7 @@ export default class ContactListScreen extends React.Component {
 							containerStyle={{}}
 							style={{ backgroundColor: '#5067FF' }}
 							position="bottomRight"
-							onPress={() => this.props.navigation.navigate('ContactCreateScreen', { navigateTo: 'ContactListScreen' })}
+							onPress={() => this.props.navigation.navigate('ProjectCreateScreen')}
 						>
 							<Icon name="create" />
 						</Fab>
